requirements.txt
streamlit
numpy
pandas
SimpleITK
pydicom
pynrrd
matplotlib
pillow
scikit-image
scipy